requirements.txt
python-whois
termcolor
requests